Move tile code to it's own file.
