+++ includes/facebook-integration.php
@@ -46,9 +46,6 @@ class Disciple_Tools_Facebook_Integration {
         $this->namespace = $this->context . "/v" . intval( $this->version );
         add_action( 'rest_api_init', [ $this, 'add_api_routes' ] );
         add_action( 'wp_ajax_dt-facebook-notice-dismiss', [ $this, 'dismiss_error' ] );
-        add_filter( "dt_custom_fields_settings", [ $this, "dt_facebook_fields" ], 1, 2 );
-        add_filter( "dt_details_additional_section_ids", [ $this, "dt_facebook_declare_section_id" ], 999, 2 );
-        add_action( "dt_details_additional_section", [ $this, "dt_facebook_add_section" ] );
         add_action( "dt_async_dt_conversation_update", [ $this, "get_conversation_update" ], 10, 2 );
         add_action( "dt_async_dt_facebook_all_conversations", [ $this, "get_conversations_with_pagination" ], 10, 4 );
         add_filter( "dt_contact_duplicate_fields_to_check", [ $this, "add_duplicate_check_field" ] );
@@ -97,166 +94,6 @@ class Disciple_Tools_Facebook_Integration {
             ]
         );
     }
-
-
-    public static function dt_facebook_fields( array $fields, string $post_type = "" ) {
-        //check if we are dealing with a contact
-        if ( $post_type === "contacts" ) {
-            //check if the language field is already set
-            if ( !isset( $fields["facebook_data"] ) ) {
-                //define the language field
-                $fields["facebook_data"] = [
-                    "name"    => __( "Facebook Ids", "dt_facebook" ),
-                    "type"    => "array",
-                    "default" => []
-                ];
-            }
-            if ( !isset( $fields["reason_closed"]["default"]["closed_from_facebook"] ) ) {
-                $fields["reason_closed"]["default"]["closed_from_facebook"] = __( "Closed from Facebook", "dt_facebook" );
-            }
-            if ( !isset( $fields["overall_status"]["default"]["from_facebook"] ) ) {
-                $fields["overall_status"]["default"]["from_facebook"] = __( "From Facebook", "dt_facebook" );
-            }
-        }
-        //don't forget to return the update fields array
-        return $fields;
-    }
-
-
-    public static function dt_facebook_declare_section_id( $sections, $post_type = "" ) {
-        //check if we are on a contact
-        if ( $post_type === "contacts" ) {
-            $contact_fields = Disciple_Tools_Contact_Post_Type::instance()->get_custom_fields_settings();
-            //check if the language field is set
-            if ( isset( $contact_fields["facebook_data"] ) ) {
-                $sections[] = "contact_facebook_data";
-            }
-            //add more section ids here if you want...
-        }
-        return $sections;
-    }
-
-
-    public static function dt_facebook_add_section( $section ) {
-        if ( $section == "contact_facebook_data" ) {
-            $contact_id    = get_the_ID();
-            $contact       = Disciple_Tools_Contacts::get_contact( $contact_id, true );
-            $facebook_data = [];
-            if ( isset( $contact["facebook_data"] ) ) {
-                $facebook_data = maybe_unserialize( $contact["facebook_data"] );
-            }
-            ?>
-            <!-- need you own css? -->
-            <style type="text/css">
-                .facebook-label {
-                    background: #ecf5fc;
-                    padding: 2px 4px;
-                    border-radius: 2px;
-                    border: 1px solid #c2e0ff;
-                }
-            </style>
-
-            <label class="section-header">
-                <?php esc_html_e( 'Facebook', 'dt_facebook' ) ?>
-            </label>
-
-
-            <?php
-            if ( isset( $facebook_data["names"] ) ) {
-                ?>
-                <div class="section-subheader">
-                    <?php esc_html_e( "Names", "dt_facebook" ) ?>
-                </div>
-                <?php
-                if ( is_array( $facebook_data["names"] ) ) {
-                    foreach ( $facebook_data["names"] as $id ) {
-                        ?>
-                        <p><?php echo esc_html( $id ) ?></p>
-                    <?php }
-                }
-            }
-            if ( isset( $facebook_data["labels"] ) ) {
-                ?>
-                <div class="section-subheader">
-                    <?php esc_html_e( "Labels", "dt_facebook" ) ?>
-                </div>
-                <p>
-                <?php foreach ( $facebook_data["labels"] as $id => $value ) {
-                    ?>
-                    <span class="facebook-label"><?php echo esc_html( $id ) ?></span>
-                <?php } ?>
-                </p>
-                <?php
-            }
-
-            if ( isset( $facebook_data["last_message_at"] ) ) {
-                $date = strtotime( $facebook_data["last_message_at"] )
-                ?>
-                <div class="section-subheader">
-                    <?php esc_html_e( "Last message at:", "dt_facebook" ) ?>
-                </div>
-                <p class="last_message_at"><?php echo esc_html( date( "Y-m-d H:m", $date ) ) ?></p>
-                <?php
-            }
-
-            if ( isset( $facebook_data["links"] ) ) {
-                ?>
-                <div class="section-subheader">
-                    <?php esc_html_e( "Conversation Links:", "dt_facebook" ) ?>
-                </div>
-                <?php
-                foreach ( $facebook_data["links"] as $link ): ?>
-                    <p class="facebook_message_links"><a href="<?php echo esc_html( 'http://facebook.com'. $link ) ?>" target="_blank"><?php echo esc_html( $link )?></a></p>
-                <?php endforeach; ?>
-                <?php
-            }
-
-
-
-
-//            foreach ( $facebook_data as $key => $value ){
-//
-            ?>
-            <!--                <div class="section-subheader">-->
-            <!--                    --><?php //echo esc_html( $key )
-            ?>
-            <!--                </div>-->
-            <!--                --><?php
-//                if ( is_array( $value )){
-//                    foreach ( $value as $id ){
-//
-            ?>
-            <!--                        <p>--><?php //echo esc_html( $id )
-            ?><!--</p>-->
-            <!--                        --><?php
-//                    }
-//                } else {
-//
-            ?>
-            <!--                    <p>--><?php //echo esc_html( $value )
-            ?><!--</p>-->
-            <!--                    --><?php
-//                }
-//            }
-
-            ?>
-
-            <script type="application/javascript">
-                //enter jquery here if you need it
-                jQuery(($) => {
-                    <?php if ( isset( $facebook_data["last_message_at"] ) ) : ?>
-                    $('.last_message_at').html(
-                        moment("<?php echo esc_html( $facebook_data["last_message_at"] ) ?>").format(
-                            "MMMM Do YYYY, h:mm:ss a"))
-                    <?php endif; ?>
-                })
-            </script>
-            <?php
-        }
-
-        //add more sections here if you want...
-    }
-
 
     /**
      * Admin notice
